refactor: Categorized response types and applied changes to related files
