+++ src/apis/responses/global.ts
@@ -1,0 +1,5 @@
+// Response types for error cases
+
+export interface ErrorResponseType {
+  message: string;
+}
