+++ src/apis/responses/member.ts
@@ -1,0 +1,12 @@
+import { DebateTable } from '../../type/type';
+
+// POST "/api/member"
+export interface PostUserResponseType {
+  id: number;
+  email: string;
+}
+
+// GET /api/table
+export interface GetDebateTableListResponseType {
+  tables: DebateTable[];
+}
