+++ src/hooks/mutations/useAddTable.ts
@@ -1,7 +1,7 @@
 import { useMutation } from '@tanstack/react-query';
 import { postParliamentaryDebateTable } from '../../apis/apis/parliamentary';
 import { DetailDebateInfo, TimeBoxInfo } from '../../type/type';
-import { PostDebateTableResponseType } from '../../apis/responseTypes';
+import { PostDebateTableResponseType } from '../../apis/responses/parliamentary';
 
 interface UseAddTableParams {
   info: DetailDebateInfo;
