+++ src/hooks/mutations/usePostUser.ts
@@ -1,5 +1,5 @@
 import { useMutation } from '@tanstack/react-query';
-import { PostUserResponseType } from '../../apis/responseTypes';
+import { PostUserResponseType } from '../../apis/responses/member';
 import { postUser } from '../../apis/apis/member';
 
 export function usePostUser(onSuccess: (data: PostUserResponseType) => void) {
